102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,9 +8,9 @@
 int main(void)
 {
 	int fn;
-	int previous;
-	int current;
-	int next;
+	long previous;
+	long current;
+	long next;
 	int i;
 
 	previous = 1;
